+++ app/src/main/java/com/fake/soundremote/ui/events/EventsScreen.kt
@@ -39,9 +39,9 @@ import androidx.compose.ui.unit.dp
 import com.fake.soundremote.R
 import com.fake.soundremote.data.Event
 import com.fake.soundremote.ui.keystrokeselectdialog.KeystrokeSelectDialog
-import com.fake.soundremote.ui.util.ListItemHeadline
-import com.fake.soundremote.ui.util.ListItemSupport
-import com.fake.soundremote.ui.util.NavigateUpButton
+import com.fake.soundremote.ui.components.ListItemHeadline
+import com.fake.soundremote.ui.components.ListItemSupport
+import com.fake.soundremote.ui.components.NavigateUpButton
 import com.fake.soundremote.util.showAppInfo
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.PermissionState
